Fix escape sequence DeprecationWarnings
In many scenarios, this just requires using raw strings or properly
escaping the values within the string. For `evaluate_terraform`, it's a
little more complex since the errors actually get emitted when the code
`eval`s the strings from the template. Properly escaping the strings
prior to `eval`ing and then still returning the input string if `eval`
fails should be sufficient to prevent such issues